requirements.txt
numpy==1.26.4
pandas==2.1.4
tokenizers==0.19.1
transformers
uvicorn==0.34.3
scikit-learn==1.6.0
fastapi==0.115.12
requests==2.32.2
matplotlib==3.9.3
xgboost==2.1.4
ibm_db_sa
sqlalchemy
#llvmlite
shap==0.47.0
fastapi[standard]
pypmml
sklearn_pandas
jupyterlab
notebook
reportlab
requests
sklearn2pmml
#pdfplumber
tqdm
beautifulsoup4
ibm-watsonx-ai
elasticsearch==8.13.0
langchain
#sentence-transformers
langchain_community
pymilvus
joblib
urllib3
certifi
pydantic
PyPDF2
opensearch-py
python-docx
#torch
pymongo
